Create hlsCurrentListners.pl
Initial commit. 
This program will look at the log and output a time stamp
and the number of listeners. The first listener column 
includes session info where available, the second is the number
of uniqie IPs.

Designed to be called from Cron once a minute.
ToDo: demonize, output to configurable logfiles